[process_modules] moved all process_modules to new directory and fixed some imports
