+++ src/pycram/process_modules/donbot_process_modules.py
@@ -1,16 +1,17 @@
+import pycram.helper_deprecated as helper_deprecated
 from pycram.robot_description import InitializedRobotDescription as robot_description
-from pycram.process_module import ProcessModule
+from pycram.process_modules.process_module import ProcessModule
 from pycram.bullet_world import BulletWorld
-from pycram.helper import _apply_ik
+from pycram.local_transformer import local_transformer
+import pycram.helper as helper
 import pycram.bullet_world_reasoning as btr
 import pybullet as p
-import logging
 import time
 
 
 def _park_arms(arm):
     """
-    Defines the joint poses for the parking positions of the arm of HSR and applies them to the
+    Defines the joint poses for the parking positions of the arm of Donbot and applies them to the
     in the BulletWorld defined robot.
     :return: None
     """
@@ -21,7 +22,7 @@ def _park_arms(arm):
             robot.set_joint_state(joint, pose)
 
 
-class HSRNavigation(ProcessModule):
+class DonbotNavigation(ProcessModule):
     """
     The process module to move the robot from one position to another.
     """
@@ -30,10 +31,15 @@ class HSRNavigation(ProcessModule):
         solution = desig.reference()
         if solution['cmd'] == 'navigate':
             robot = BulletWorld.robot
+            # Reset odom joints to zero
+            for joint_name in robot_description.i.odom_joints:
+                robot.set_joint_state(joint_name, 0.0)
+            # Set actual goal pose
             robot.set_position_and_orientation(solution['target'], solution['orientation'])
+            time.sleep(0.5)
+            local_transformer.update_from_btr()
 
-
-class HSRPickUp(ProcessModule):
+class DonbotPickUp(ProcessModule):
     """
     This process module is for picking up a given object.
     The object has to be reachable for this process module to succeed.
@@ -47,12 +53,12 @@ class HSRPickUp(ProcessModule):
             target = object.get_position()
             inv = p.calculateInverseKinematics(robot.id, robot.get_link_id(solution['gripper']), target,
                                                maxNumIterations=100)
-            _apply_ik(robot, inv)
+            helper._apply_ik(robot, inv)
             robot.attach(object, solution['gripper'])
             time.sleep(0.5)
 
 
-class HSRPlace(ProcessModule):
+class DonbotPlace(ProcessModule):
     """
     This process module places an object at the given position in world coordinate frame.
     """
@@ -64,12 +70,12 @@ class HSRPlace(ProcessModule):
             robot = BulletWorld.robot
             inv = p.calculateInverseKinematics(robot.id, robot.get_link_id(solution['gripper']), solution['target'],
                                                maxNumIterations=100)
-            _apply_ik(robot, inv)
+            helper._apply_ik(robot, inv)
             robot.detach(object)
             time.sleep(0.5)
 
 
-class HSRAccessing(ProcessModule):
+class DonbotAccessing(ProcessModule):
     """
     This process module responsible for opening drawers to access the objects inside. This works by firstly moving
     the end effector to the handle of the drawer. Next, the end effector is moved the respective distance to the back.
@@ -88,17 +94,17 @@ class HSRAccessing(ProcessModule):
             dis = solution['distance']
             inv = p.calculateInverseKinematics(robot.id, robot.get_link_id(gripper),
                                                kitchen.get_link_position(drawer_handle))
-            _apply_ik(robot, inv)
+            helper._apply_ik(robot, inv)
             time.sleep(0.2)
             han_pose = kitchen.get_link_position(drawer_handle)
             new_p = [han_pose[0] - dis, han_pose[1], han_pose[2]]
             inv = p.calculateInverseKinematics(robot.id, robot.get_link_id(gripper), new_p)
-            _apply_ik(robot, inv)
+            helper._apply_ik(robot, inv)
             kitchen.set_joint_state(drawer_joint, 0.3)
             time.sleep(0.5)
 
 
-class HSRParkArms(ProcessModule):
+class DonbotParkArms(ProcessModule):
     """
     This process module is for moving the arms in a parking position.
     It is currently not used.
@@ -110,7 +116,7 @@ class HSRParkArms(ProcessModule):
             _park_arms()
 
 
-class HSRMoveHead(ProcessModule):
+class DonbotMoveHead(ProcessModule):
     """
     This process module moves the head to look at a specific point in the world coordinate frame.
     This point can either be a position or an object.
@@ -119,16 +125,36 @@ class HSRMoveHead(ProcessModule):
     def _execute(self, desig):
         solutions = desig.reference()
         if solutions['cmd'] == 'looking':
-            target = solutions['target']
-            if target is 'forward' or target is 'down':
-                robot = BulletWorld.robot
-                for joint, state in robot_description.i.get_static_joint_chain("neck", target).items():
-                    robot.set_joint_state(joint, state)
+            robot = BulletWorld.robot
+            neck_base_frame = local_transformer.projection_namespace + '/' + robot_description.i.chains["neck"].base_link if \
+                local_transformer.projection_namespace else \
+                robot_description.i.chains["neck"].base_link
+            if type(solutions['target']) is str:
+                target = local_transformer.projection_namespace + '/' + solutions['target'] if \
+                    local_transformer.projection_namespace else \
+                    solutions['target']
+                pose_in_neck_base = local_transformer.tf_transform(neck_base_frame, target)
+            elif helper_deprecated.is_list_pose(solutions['target']) or helper_deprecated.is_list_position(solutions['target']):
+                pose = helper_deprecated.ensure_pose(solutions['target'])
+                pose_in_neck_base = local_transformer.tf_pose_transform(local_transformer.map_frame, neck_base_frame, pose)
+
+            vector = pose_in_neck_base[0]
+            # +x as forward
+            # +y as left
+            # +z as up
+            x = vector[0]
+            y = vector[1]
+            z = vector[2]
+            conf = None
+            if y > 0:
+                conf = "left"
             else:
-                logging.error("There is no target position defined with the target %s.", target)
+                conf = "right"
+            for joint, state in robot_description.i.get_static_joint_chain("neck", conf).items():
+                robot.set_joint_state(joint, state)
 
 
-class HSRMoveGripper(ProcessModule):
+class DonbotMoveGripper(ProcessModule):
     """
     This process module controls the gripper of the robot. They can either be opened or closed.
     Furthermore, it can only moved one gripper at a time.
@@ -141,11 +167,12 @@ class HSRMoveGripper(ProcessModule):
             gripper = solution['gripper']
             motion = solution['motion']
             for joint, state in robot_description.i.get_static_gripper_chain(gripper, motion).items():
+                # TODO: Test this, add gripper-opening/-closing to the demo.py
                 robot.set_joint_state(joint, state)
             time.sleep(0.5)
 
 
-class HSRDetecting(ProcessModule):
+class DonbotDetecting(ProcessModule):
     """
     This process module tries to detect an object with the given type. To be detected the object has to be in
     the field of view of the robot.
@@ -165,7 +192,7 @@ class HSRDetecting(ProcessModule):
                     return obj
 
 
-class HSRMoveTCP(ProcessModule):
+class DonbotMoveTCP(ProcessModule):
     """
     This process moves the tool center point of either the right or the left arm.
     """
@@ -177,11 +204,11 @@ class HSRMoveTCP(ProcessModule):
             gripper = solution['gripper']
             robot = BulletWorld.robot
             inv = p.calculateInverseKinematics(robot.id, robot.get_link_id(gripper), target)
-            _apply_ik(robot, inv)
+            helper._apply_ik(robot, inv)
             time.sleep(0.5)
 
 
-class HSRMoveJoints(ProcessModule):
+class DonbotMoveJoints(ProcessModule):
     """
     This process modules moves the joints of either the right or the left arm. The joint states can be given as
     list that should be applied or a pre-defined position can be used, such as "parking"
@@ -202,7 +229,7 @@ class HSRMoveJoints(ProcessModule):
             time.sleep(0.5)
 
 
-class HSRWorldStateDetecting(ProcessModule):
+class DonbotWorldStateDetecting(ProcessModule):
     """
     This process module detectes an object even if it is not in the field of view of the robot.
     """
@@ -213,17 +240,16 @@ class HSRWorldStateDetecting(ProcessModule):
             obj_type = solution['object_type']
             return list(filter(lambda obj: obj.type == obj_type, BulletWorld.current_bullet_world.objects))[0]
 
+DonbotProcessModulesSimulated = {'moving' : DonbotNavigation(),
+                              'pick-up' : DonbotPickUp(),
+                              'place' : DonbotPlace(),
+                              'accessing' : DonbotAccessing(),
+                              'looking' : DonbotMoveHead(),
+                              'opening_gripper' : DonbotMoveGripper(),
+                              'closing_gripper' : DonbotMoveGripper(),
+                              'detecting' : DonbotDetecting(),
+                              'move-tcp' : DonbotMoveTCP(),
+                              'move-arm-joints' : DonbotMoveJoints(),
+                              'world-state-detecting' : DonbotWorldStateDetecting()}
 
-HSRProcessModulesSimulated = {'moving' : HSRNavigation(),
-                              'pick-up' : HSRPickUp(),
-                              'place' : HSRPlace(),
-                              'accessing' : HSRAccessing(),
-                              'looking' : HSRMoveHead(),
-                              'opening_gripper' : HSRMoveGripper(),
-                              'closing_gripper' : HSRMoveGripper(),
-                              'detecting' : HSRDetecting(),
-                              'move-tcp' : HSRMoveTCP(),
-                              'move-arm-joints' : HSRMoveJoints(),
-                              'world-state-detecting' : HSRWorldStateDetecting()}
-
-HSRProcessModulesReal = {}
+DonbotProcessModulesReal = {}
