+++ src/pycram/process_modules/hsr_process_modules.py
@@ -1,18 +1,16 @@
-import pycram.helper_deprecated as helper_deprecated
 from pycram.robot_description import InitializedRobotDescription as robot_description
-from pycram.process_module import ProcessModule
+from pycram.process_modules.process_module import ProcessModule
 from pycram.bullet_world import BulletWorld
-from pycram.local_transformer import local_transformer
-import pycram.helper as helper
+from pycram.helper import _apply_ik
 import pycram.bullet_world_reasoning as btr
 import pybullet as p
-from rospy import logerr
+import logging
 import time
 
 
 def _park_arms(arm):
     """
-    Defines the joint poses for the parking positions of the arm of Donbot and applies them to the
+    Defines the joint poses for the parking positions of the arm of HSR and applies them to the
     in the BulletWorld defined robot.
     :return: None
     """
@@ -23,7 +21,7 @@ def _park_arms(arm):
             robot.set_joint_state(joint, pose)
 
 
-class DonbotNavigation(ProcessModule):
+class HSRNavigation(ProcessModule):
     """
     The process module to move the robot from one position to another.
     """
@@ -32,15 +30,10 @@ class DonbotNavigation(ProcessModule):
         solution = desig.reference()
         if solution['cmd'] == 'navigate':
             robot = BulletWorld.robot
-            # Reset odom joints to zero
-            for joint_name in robot_description.i.odom_joints:
-                robot.set_joint_state(joint_name, 0.0)
-            # Set actual goal pose
             robot.set_position_and_orientation(solution['target'], solution['orientation'])
-            time.sleep(0.5)
-            local_transformer.update_from_btr()
 
-class DonbotPickUp(ProcessModule):
+
+class HSRPickUp(ProcessModule):
     """
     This process module is for picking up a given object.
     The object has to be reachable for this process module to succeed.
@@ -54,12 +47,12 @@ class DonbotPickUp(ProcessModule):
             target = object.get_position()
             inv = p.calculateInverseKinematics(robot.id, robot.get_link_id(solution['gripper']), target,
                                                maxNumIterations=100)
-            helper._apply_ik(robot, inv)
+            _apply_ik(robot, inv)
             robot.attach(object, solution['gripper'])
             time.sleep(0.5)
 
 
-class DonbotPlace(ProcessModule):
+class HSRPlace(ProcessModule):
     """
     This process module places an object at the given position in world coordinate frame.
     """
@@ -71,12 +64,12 @@ class DonbotPlace(ProcessModule):
             robot = BulletWorld.robot
             inv = p.calculateInverseKinematics(robot.id, robot.get_link_id(solution['gripper']), solution['target'],
                                                maxNumIterations=100)
-            helper._apply_ik(robot, inv)
+            _apply_ik(robot, inv)
             robot.detach(object)
             time.sleep(0.5)
 
 
-class DonbotAccessing(ProcessModule):
+class HSRAccessing(ProcessModule):
     """
     This process module responsible for opening drawers to access the objects inside. This works by firstly moving
     the end effector to the handle of the drawer. Next, the end effector is moved the respective distance to the back.
@@ -95,17 +88,17 @@ class DonbotAccessing(ProcessModule):
             dis = solution['distance']
             inv = p.calculateInverseKinematics(robot.id, robot.get_link_id(gripper),
                                                kitchen.get_link_position(drawer_handle))
-            helper._apply_ik(robot, inv)
+            _apply_ik(robot, inv)
             time.sleep(0.2)
             han_pose = kitchen.get_link_position(drawer_handle)
             new_p = [han_pose[0] - dis, han_pose[1], han_pose[2]]
             inv = p.calculateInverseKinematics(robot.id, robot.get_link_id(gripper), new_p)
-            helper._apply_ik(robot, inv)
+            _apply_ik(robot, inv)
             kitchen.set_joint_state(drawer_joint, 0.3)
             time.sleep(0.5)
 
 
-class DonbotParkArms(ProcessModule):
+class HSRParkArms(ProcessModule):
     """
     This process module is for moving the arms in a parking position.
     It is currently not used.
@@ -117,7 +110,7 @@ class DonbotParkArms(ProcessModule):
             _park_arms()
 
 
-class DonbotMoveHead(ProcessModule):
+class HSRMoveHead(ProcessModule):
     """
     This process module moves the head to look at a specific point in the world coordinate frame.
     This point can either be a position or an object.
@@ -126,36 +119,16 @@ class DonbotMoveHead(ProcessModule):
     def _execute(self, desig):
         solutions = desig.reference()
         if solutions['cmd'] == 'looking':
-            robot = BulletWorld.robot
-            neck_base_frame = local_transformer.projection_namespace + '/' + robot_description.i.chains["neck"].base_link if \
-                local_transformer.projection_namespace else \
-                robot_description.i.chains["neck"].base_link
-            if type(solutions['target']) is str:
-                target = local_transformer.projection_namespace + '/' + solutions['target'] if \
-                    local_transformer.projection_namespace else \
-                    solutions['target']
-                pose_in_neck_base = local_transformer.tf_transform(neck_base_frame, target)
-            elif helper_deprecated.is_list_pose(solutions['target']) or helper_deprecated.is_list_position(solutions['target']):
-                pose = helper_deprecated.ensure_pose(solutions['target'])
-                pose_in_neck_base = local_transformer.tf_pose_transform(local_transformer.map_frame, neck_base_frame, pose)
-
-            vector = pose_in_neck_base[0]
-            # +x as forward
-            # +y as left
-            # +z as up
-            x = vector[0]
-            y = vector[1]
-            z = vector[2]
-            conf = None
-            if y > 0:
-                conf = "left"
+            target = solutions['target']
+            if target is 'forward' or target is 'down':
+                robot = BulletWorld.robot
+                for joint, state in robot_description.i.get_static_joint_chain("neck", target).items():
+                    robot.set_joint_state(joint, state)
             else:
-                conf = "right"
-            for joint, state in robot_description.i.get_static_joint_chain("neck", conf).items():
-                robot.set_joint_state(joint, state)
+                logging.error("There is no target position defined with the target %s.", target)
 
 
-class DonbotMoveGripper(ProcessModule):
+class HSRMoveGripper(ProcessModule):
     """
     This process module controls the gripper of the robot. They can either be opened or closed.
     Furthermore, it can only moved one gripper at a time.
@@ -168,12 +141,11 @@ class DonbotMoveGripper(ProcessModule):
             gripper = solution['gripper']
             motion = solution['motion']
             for joint, state in robot_description.i.get_static_gripper_chain(gripper, motion).items():
-                # TODO: Test this, add gripper-opening/-closing to the demo.py
                 robot.set_joint_state(joint, state)
             time.sleep(0.5)
 
 
-class DonbotDetecting(ProcessModule):
+class HSRDetecting(ProcessModule):
     """
     This process module tries to detect an object with the given type. To be detected the object has to be in
     the field of view of the robot.
@@ -193,7 +165,7 @@ class DonbotDetecting(ProcessModule):
                     return obj
 
 
-class DonbotMoveTCP(ProcessModule):
+class HSRMoveTCP(ProcessModule):
     """
     This process moves the tool center point of either the right or the left arm.
     """
@@ -205,11 +177,11 @@ class DonbotMoveTCP(ProcessModule):
             gripper = solution['gripper']
             robot = BulletWorld.robot
             inv = p.calculateInverseKinematics(robot.id, robot.get_link_id(gripper), target)
-            helper._apply_ik(robot, inv)
+            _apply_ik(robot, inv)
             time.sleep(0.5)
 
 
-class DonbotMoveJoints(ProcessModule):
+class HSRMoveJoints(ProcessModule):
     """
     This process modules moves the joints of either the right or the left arm. The joint states can be given as
     list that should be applied or a pre-defined position can be used, such as "parking"
@@ -230,7 +202,7 @@ class DonbotMoveJoints(ProcessModule):
             time.sleep(0.5)
 
 
-class DonbotWorldStateDetecting(ProcessModule):
+class HSRWorldStateDetecting(ProcessModule):
     """
     This process module detectes an object even if it is not in the field of view of the robot.
     """
@@ -241,16 +213,17 @@ class DonbotWorldStateDetecting(ProcessModule):
             obj_type = solution['object_type']
             return list(filter(lambda obj: obj.type == obj_type, BulletWorld.current_bullet_world.objects))[0]
 
-DonbotProcessModulesSimulated = {'moving' : DonbotNavigation(),
-                              'pick-up' : DonbotPickUp(),
-                              'place' : DonbotPlace(),
-                              'accessing' : DonbotAccessing(),
-                              'looking' : DonbotMoveHead(),
-                              'opening_gripper' : DonbotMoveGripper(),
-                              'closing_gripper' : DonbotMoveGripper(),
-                              'detecting' : DonbotDetecting(),
-                              'move-tcp' : DonbotMoveTCP(),
-                              'move-arm-joints' : DonbotMoveJoints(),
-                              'world-state-detecting' : DonbotWorldStateDetecting()}
 
-DonbotProcessModulesReal = {}
+HSRProcessModulesSimulated = {'moving' : HSRNavigation(),
+                              'pick-up' : HSRPickUp(),
+                              'place' : HSRPlace(),
+                              'accessing' : HSRAccessing(),
+                              'looking' : HSRMoveHead(),
+                              'opening_gripper' : HSRMoveGripper(),
+                              'closing_gripper' : HSRMoveGripper(),
+                              'detecting' : HSRDetecting(),
+                              'move-tcp' : HSRMoveTCP(),
+                              'move-arm-joints' : HSRMoveJoints(),
+                              'world-state-detecting' : HSRWorldStateDetecting()}
+
+HSRProcessModulesReal = {}
