+++ src/pycram/process_modules/pr2_process_modules.py
@@ -1,12 +1,11 @@
 from pycram.robot_description import InitializedRobotDescription as robot_description
-from pycram.process_module import ProcessModule
+from pycram.process_modules.process_module import ProcessModule
 from pycram.bullet_world import BulletWorld
 from pycram.helper import transform
 from pycram.ik import request_ik
 from pycram.helper import _transform_to_torso, _apply_ik
 from pycram.local_transformer import local_transformer
 import pycram.bullet_world_reasoning as btr
-import pybullet as p
 import numpy as np
 import time
